Fix failing migration from AM test
